* mcc/misc: use 32bit values for all TAB sizes.

+++ mcc/Methods.c
@@ -340,12 +340,12 @@ IPTR mInputTrigger(struct IClass *cl, Object *obj, UNUSED Msg msg)
       else
       {
         struct pos_info pos;
-        ULONG flow;
+        LONG flow;
 
         OffsetToLines(data, data->CPos_X, data->actualline, &pos);
         flow = FlowSpace(data, data->actualline->line.Flow, &data->actualline->line.Contents[pos.bytes]);
 
-        if(MouseX <= (LONG)(_mleft(obj)+flow+TextLengthNew(&data->tmprp, &data->actualline->line.Contents[pos.bytes], pos.extra-pos.bytes-1, data->TabSizePixels)))
+        if(MouseX <= _mleft(obj)+flow+TextLengthNew(&data->tmprp, &data->actualline->line.Contents[pos.bytes], pos.extra-pos.bytes-1, data->TabSizePixels))
         {
           if(data->selectmode == 1)
           {
